Improvement CardSliderVM
